Moved serializers to reporting package
