+++ cli/pom.xml
@@ -20,7 +20,6 @@
         <dependency>
             <groupId>de.jplag</groupId>
             <artifactId>reporting</artifactId>
-            <version>${revision}</version>
         </dependency>
 
         <!-- Languages -->
