+++ coverage-report/pom.xml
@@ -36,7 +36,6 @@
         <dependency>
             <groupId>de.jplag</groupId>
             <artifactId>reporting</artifactId>
-            <version>${revision}</version>
         </dependency>
 
         <!-- Languages -->
