+++ endtoend-testing/pom.xml
@@ -19,6 +19,10 @@
         </dependency>
         <dependency>
             <groupId>de.jplag</groupId>
+            <artifactId>reporting</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>de.jplag</groupId>
             <artifactId>language-testutils</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
