+++ pom.xml
@@ -359,6 +359,11 @@
             </dependency>
             <dependency>
                 <groupId>de.jplag</groupId>
+                <artifactId>reporting</artifactId>
+                <version>${revision}</version>
+            </dependency>
+            <dependency>
+                <groupId>de.jplag</groupId>
                 <artifactId>language-api</artifactId>
                 <version>${revision}</version>
             </dependency>
